[!] Initial commit..
Input digital pin done.

But I'm not happy with interface.

We're storing more information than we need for work. And of course
this redundant information can be contradictory

: For work we need byte address and bit offset. That's three bytes.
But Init() accepts pin number. Because no one fucking going to
provide byte addresses and bit offsets.

Problem is that we wanna be nice, we should provide interface to
give those pin number back. Or should we?

Similar shit for saturation bit. We don't need to store anything
for work but are storing boolean value just to be nice and return
it back if asked.

--

I gonna strip this design. But maybe I'll regret this decision,
so I can revert to this commit.